* amandad-src/amandad.c: Do not free as * recover-src/extract_list.c: Check state_filename != NULL before   unlink
git-svn-id: https://svn.code.sf.net/p/amanda/code/amanda/trunk@6867 a8d146d6-cc15-0410-8900-af154a0219e0

+++ recover-src/extract_list.c
@@ -3126,9 +3126,11 @@ read_amidxtaped_state(
 
 	aclose(ctl_state->fd);
 	if (ctl_state->bytes_read == 0) {
-	    unlink(state_filename);
-	    g_free(state_filename);
-	    state_filename = NULL;
+	    if (state_filename) {
+		unlink(state_filename);
+		g_free(state_filename);
+		state_filename = NULL;
+	    }
 	}
 	ctl_state->state_done = TRUE;
 	if (am_has_feature(tapesrv_features, fe_amrecover_state_done)) {
